deck_service/requirements.txt
django
djangorestframework
psycopg2-binary
python-decouple
drf-spectacular
drf-spectacular-sidecar
kafka-python
requests>=2.25.0
whitenoise>=6.0.0 